% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{add_datetime_columns}
\alias{add_datetime_columns}
\title{add_datetime_columns
Adds columns that allow for easier temporal deliniation
add date columns with lubridate for use in querying by date}
\usage{
add_datetime_columns(detection_dataframe)
}
\description{
add_datetime_columns
Adds columns that allow for easier temporal deliniation
add date columns with lubridate for use in querying by date
}
