% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_in_period_resolution.R
\name{date_time_columns_selector}
\alias{date_time_columns_selector}
\title{gets column indexes of specified resolution to allow grouping for the
function detected_in_period_resolution}
\usage{
date_time_columns_selector(resolution, data)
}
\arguments{
\item{resolution}{The resolution you want to view the data divided into
specified as a character string of one of the following: "year", "month",
"week", "day", or "hour"}

\item{data}{The dataset you are wanting to group within.}
}
\value{
Returns list of column names that allow grouping by resolution
}
\description{
gets column indexes of specified resolution to allow grouping for the
function detected_in_period_resolution
}
