% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_in_period_resolution.R
\name{detected_in_period_resolution}
\alias{detected_in_period_resolution}
\title{detected_in_period_resolution.R}
\usage{
detected_in_period_resolution(data_set, start_date = NULL,
  end_date = NULL, known_tags, detection_resolution)
}
\arguments{
\item{start_date}{Enter as character string "Month-Day-Year"}

\item{end_date}{Enter as character string "Month-Day-Year"}

\item{resolution}{The resolution you want to view the data divided into,
specified as a character string of one of the folling: "year", "month", "week", "day"
detected_in_period()}
}
\description{
This function allows you to detect which specimens from a list were detected in
a specified period divided by a specified resolution
}
