% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first_last}
\alias{first_last}
\title{first_last}
\usage{
first_last(data_set, start_date = NULL, end_date = NULL)
}
\arguments{
\item{data_set}{Detection data. Use \code{\link{get_detection_data}} to
import data first.}

\item{start_date}{Optional. Use to specify beginning of observation range.}

\item{end_date}{Optional. Use to specify end of observation range.}
}
\value{
Returns data frame with variables: bleh, ah, sing, crazy_hair
}
\description{
first_last
}
