% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_encounter_history.R
\name{generate_data_for_survival_analysis}
\alias{generate_data_for_survival_analysis}
\title{generate data for survival analysis}
\usage{
generate_data_for_survival_analysis(data_set, tags, dates)
}
\arguments{
\item{data_set}{Data read in using
\code{\link{get_detection_data}}.}

\item{tags}{Dataframe of tags to be examined}

\item{dates}{Character vector containing the dates at which you want to break
the analysis into. Use "mm-dd-yyyy" format when entering dates.}
}
\description{
Used for preparing data for survival analysis using MARK software.
}
