% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{get_tag_list}
\alias{get_tag_list}
\title{Read in tags}
\usage{
get_tag_list(tags_text_file)
}
\arguments{
\item{tags_text_file}{.csv file containing tags in the the second column.}
}
\value{
Returns dataframe with one column named "tags" for use with other
  functions in package.
}
\description{
Reads in tags from .csv file. Expects row numbers in first column that are
removed. Skips first line as column name.
}
