% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_from_database.R
\name{get_tags_from_db_all_species}
\alias{get_tags_from_db_all_species}
\title{tags_from_database.r
function to extract tags from databas}
\usage{
get_tags_from_db_all_species(database_path)
}
\arguments{
\item{database_path}{The path to the database youwant to use, use forward slashes

Function to extract list of tags from database, either for a particular
species or for all species
must be running 32 bit R for this to function}
}
\description{
tags_from_database.r
function to extract tags from databas
}
