% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_resolution.R
\name{group_by_resolution}
\alias{group_by_resolution}
\title{group_by_resolution}
\usage{
group_by_resolution(data, resolution)
}
\arguments{
\item{data}{-The dataset that you want to get the resolution grouping of}

\item{resolution}{-The resolution you want to group the data by}
}
\description{
Groups data frame by resolution, orders by date
}
