% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_direction.R
\name{movement_detection}
\alias{movement_detection}
\title{Movement direction detection}
\usage{
movement_detection(dataset)
}
\arguments{
\item{dataset}{The data set containing the observations}
}
\value{
Returns a dataframe containing the tag, detection time/date info,
  direction of travel, and which detector interval made the observation.
}
\description{
Returns data on directional movement across detection array.
}
