% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\docType{class}
\name{myDate-class}
\alias{myDate-class}
\title{Class to import datetime from character variable in original dataset in correct format.}
\description{
Class to import datetime from character variable in original dataset in correct format.
}
