% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number_species_by_resolution.R
\name{plot_by_species_resolution}
\alias{plot_by_species_resolution}
\title{plot_number_species_by_resolution.r}
\usage{
plot_by_species_resolution(data_set, start_date = NULL,
  end_date = NULL, tags, detection_resolution)
}
\description{
plot_number_species_by_resolution.r
}
