% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartEndToDates.R
\name{StartEndToDates}
\alias{StartEndToDates}
\title{StartEndToDates}
\usage{
StartEndToDates(start_date, end_date)
}
\arguments{
\item{start_date}{Enter as character string "Month-Day-Year"}

\item{end_date}{Enter as character string "Month-Day-Year"}
}
\description{
StartEndToDates
}
