% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetbydaterange.R
\name{subset_by_date_range}
\alias{subset_by_date_range}
\title{subset_by_date_range}
\usage{
subset_by_date_range(data_set, date_col = "detected_at", start_date,
  end_date, na.rm = FALSE)
}
\arguments{
\item{data_set}{The data set to get data from}

\item{date_col}{The column in the data set storing the date
of observation}

\item{start_date}{Enter as character string "mm-dd-yyyy"}

\item{end_date}{Enter as character string "mm-dd-yyyy"}
}
\value{
A data frame subsetted to observations only inclusively
within specified range, but including all original columns
}
\description{
subset_by_date_range
}
