% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_from_database.R
\name{tags_from_db}
\alias{tags_from_db}
\title{tags_from_db}
\usage{
tags_from_db(database_path)
}
\arguments{
\item{database_path}{The path to the database you want to use, use forward
slashes to seperate directories. Follows GCMRCC database format and
extracts the 'PITTAG' and 'SPECIES_CODE' from the 'FISH_T_SPECIMEN'
table in the database.}
}
\value{
Returns data frame of two variables. 'FISH_T_SPECIMEN.PITTAG' is
  returned as "tag" in the new dataframe and 'FISH_T_SPECIMEN.SPECIES_CODE'
  is returned as "species" so that the variable names match those found
  throughout the 'pitDataR' package.
}
\description{
function to extract tags along with their corresponding species from a
database following the GCMRC convention
}
\section{Warning}{
 Must use 32-bit R version to function.
}

