% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{tags_with_species_code}
\alias{tags_with_species_code}
\title{tags_with_species_code}
\usage{
tags_with_species_code(tags_text_file)
}
\arguments{
\item{tags_text_file}{A csv file containing tags and species code. Expects
file to how column of row numbers but then removes them. Skips line
containing columnames.}
}
\value{
Returns dataframe with two columns "tag" and "species" to be used
  with other functions in this package.
}
\description{
Gets tag numbers and species code from csv file
}
