% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueFalseDetectionInRange.R
\name{TrueFalseDetectionInRange}
\alias{TrueFalseDetectionInRange}
\title{TrueFalseDetectionInRange}
\usage{
TrueFalseDetectionInRange(data_set, start_date, end_date, tag_list)
}
\description{
TrueFalseDetectionInRange
}
