% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_tags.R
\name{add_all_tags}
\alias{add_all_tags}
\title{add_all_tags.R
This function creates a data frame from all of the possible tags that matches
the format of the detection data, then joins it with the detection data
to produce a table to be fed to MARK to determine capture probabilities.}
\usage{
add_all_tags(detection, tags)
}
\description{
add_all_tags.R
This function creates a data frame from all of the possible tags that matches
the format of the detection data, then joins it with the detection data
to produce a table to be fed to MARK to determine capture probabilities.
}
