% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_in_period.R,
%   R/detected_in_period_from_old.R
\name{detected_in_period}
\alias{detected_in_period}
\title{detected_in_period.R
check over this functionality
This is a method to determine if there was a detection of a particular tag
in a specied period and return a bolean of detections}
\usage{
detected_in_period(data, start_date = min(data[["detected_at"]]),
  end_date = max(data[["detected_at"]]), known_tags)

detected_in_period(data, start_date = min(data[["detected_at"]]),
  end_date = max(data[["detected_at"]]), known_tags)
}
\description{
detected_in_period.R
check over this functionality
This is a method to determine if there was a detection of a particular tag
in a specied period and return a bolean of detections

detected_in_period.R
This is a method to determine if there was a detection of a particular tag
in a specied period and return a bolean of detections
}
