% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detected_in_period.R
\name{detected_in_period}
\alias{detected_in_period}
\title{detected_in_period.R This is a method to determine if there was a detection
of a particular tag in a specied period and return a bolean of detections}
\usage{
detected_in_period(data_set, start_date = NULL, end_date = NULL,
  known_tags)
}
\arguments{
\item{start_date}{Enter as character string "MM-DD-YYYY"}

\item{end_date}{Enter as character string "MM-DD-YYYY"}

\item{known_tags}{Tags either imported using\code{\link{tags_from_db}}.}

\item{detection_dataframe}{Data read in using
\code{\link{get_detection_data}}.}
}
\description{
detected_in_period.R This is a method to determine if there was a detection
of a particular tag in a specied period and return a bolean of detections
}
