<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_for_survival_analysis.R
\name{generate_data_for_survival_analysis}
\alias{generate_data_for_survival_analysis}
\title{generate data for survival analysis}
\usage{
generate_data_for_survival_analysis(data_set, tags, dates)
}
\arguments{
\item{data_set}{Data read in using \code{\link{get_detection_data}}.}

\item{tags}{Dataframe of tags to be examined}

\item{dates}{Character vector containing the dates at which you want to break
the analysis into. Use "mm-dd-yyyy" format when entering dates.}
}
\value{
Eventually the function will return a .txt file of the data formatted
  to the detection history format used by MARK software that will allow for
  survival analysis. Currently it creates a dataframe object containing
  columns for tag, a initial detection, and each subsequent detection period.
  The detection period data is recorded as a either a 0 for no detection in
  that period or as a 1 indicating a detection in that period.
}
\description{
Used for preparing data for survival analysis using MARK software.
}
\note{
This generates data specifically for mark/recapture survivalship
  analysis. For more information see
  \href{http://www.phidot.org/software/mark/docs/book/}{\emph{Program MARK: a
  gentle introduction}}.
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_encounter_history.R
\name{generate_data_for_survival_analysis}
\alias{generate_data_for_survival_analysis}
\title{generate data for survival analysis}
\usage{
generate_data_for_survival_analysis(data_set, tags, dates)
}
\arguments{
\item{data_set}{Data read in using
\code{\link{get_detection_data}}.}

\item{tags}{Dataframe of tags to be examined}

\item{dates}{Character vector containing the dates at which you want to break
the analysis into. Use "mm-dd-yyyy" format when entering dates.}
}
\description{
Used for preparing data for survival analysis using MARK software.
}
>>>>>>> aab2ae6828121046b264b00cd695f3888df1939e
