% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{get_detection_data}
\alias{get_detection_data}
\title{Get data to work with}
\usage{
get_detection_data(detection_data)
}
\arguments{
\item{detection_data}{The containing pittag detections as directly downloaded}
}
\description{
Get data to work with
}
