% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in.R
\name{get_detection_data}
\alias{get_detection_data}
\title{get_detection_data
Reads in detection data}
\usage{
get_detection_data(data_set)
}
\arguments{
\item{data_set}{Path to data containing pittag detections as directly downloaded from Biomark's BIOlogic in .csv format. Use forward slashes to specify path.}
}
\value{
Returns dataframe with variables: tag, detected_at, reader_antenna, year, month, week, day, hour
}
\description{
get_detection_data
Reads in detection data
}
