% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_direction.R
\name{movement_detection}
\alias{movement_detection}
\title{movement_direction}
\usage{
movement_detection(dataset)
}
\arguments{
\item{dataset}{The data set containing the observations}
}
\value{
dataframe containing the tag, detection time/date info,
direction of travel and over which detector interval the observation was made
}
\description{
movement_direction
}
