% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_tags_of_interest.R
\name{only_tags_of_interest}
\alias{only_tags_of_interest}
\title{only_tags_of_interest.R
subsets all of the data to include only the values from list of tags.}
\usage{
only_tags_of_interest(data, tags)
}
\arguments{
\item{data}{the dataframe containing all detection data}

\item{tags}{the dataframe containing the tags of interest.}
}
\description{
only_tags_of_interest.R
subsets all of the data to include only the values from list of tags.
}
